Algorithm finished. Don't bother me again with that D:<

+++ apirest/src/models/users/offerers/offerers.js
@@ -153,14 +153,14 @@ module.exports = (sequelize, DataTypes) => {
       },
 
       satisfactionAVG: {
-        type: DataTypes.FLOAT(3, 2),
+        type: DataTypes.FLOAT(),
         allowNull: false,
         defaultValue: 0,
         validate: {
           isFloat: true
         }
       },
-      
+
       nOpinions: {
         type: DataTypes.INTEGER,
         allowNull: false,
@@ -178,7 +178,7 @@ module.exports = (sequelize, DataTypes) => {
         }
       },
       ratioSuccess: {
-        type: DataTypes.FLOAT(),
+        type: DataTypes.FLOAT(4, 2),
         allowNull: false,
         defaultValue: 0,
         validate: {
